# perf(charts): improve performance on GET list