--- a/superset/models/slice.py
+++ b/superset/models/slice.py
@@ -72,6 +72,15 @@ class Slice(
     perm = Column(String(1000))
     schema_perm = Column(String(1000))
     owners = relationship(security_manager.user_model, secondary=slice_user)
+    table = relationship(
+        "SqlaTable",
+        foreign_keys=[datasource_id],
+        primaryjoin="and_(Slice.datasource_id == SqlaTable.id, "
+        "Slice.datasource_type == 'table')",
+        remote_side="SqlaTable.id",
+        lazy="joined",
+    )
+
     token = ""
 
     export_fields = [
@@ -121,13 +130,15 @@ class Slice(
     @renders("datasource_url")
     def datasource_url(self) -> Optional[str]:
         # pylint: disable=no-member
-        datasource = self.datasource
-        return datasource.explore_url if datasource else None
+        if self.table:
+            return self.table.explore_url
+        return None
 
     def datasource_name_text(self) -> Optional[str]:
         # pylint: disable=no-member
-        datasource = self.datasource
-        return datasource.name if datasource else None
+        if self.table:
+            return self.table.table_name
+        return None
 
     @property
     def datasource_edit_url(self) -> Optional[str]: