### CATEGORY

Choose one

- [ ] Bug Fix
- [X] Enhancement (new features, refinement)
- [ ] Refactor
- [ ] Add tests
- [ ] Build / Development Environment
- [ ] Documentation

### SUMMARY
This API endpoint was issuing a couple of extra queries for each row. One for `ab_user` resolving created by field and another for fetching the correct datasource. Since druid support outside of SQLAlchemy is deprecated, making an outer join with SqlaTable.

The idea to optimize and avoid a query per row, is when using `@property` has a column and the method itself references a column, make sure this column is declared on `list_columns` so it's prefetched or SQLAlchemy will issue the extra query 

Local times are:
Before:
(timing) ChartRestApi.get_list.time = 130ms - 180ms

After:
(timing) ChartRestApi.get_list.time = 25ms - 50ms

Druid charts get displayed normally but if any, an extra query is issued for each one:

<img width="1149" alt="Screenshot 2020-04-30 at 15 34 17" src="https://user-images.githubusercontent.com/4025227/80723100-18c04e00-8af8-11ea-9bba-81d211918394.png">

### ADDITIONAL INFORMATION
- [ ] Has associated issue:
- [X] Changes UI
- [ ] Requires DB Migration.
- [ ] Confirm DB Migration upgrade and downgrade tested.
- [ ] Introduces new feature or API
- [ ] Removes existing feature or API

### REVIEWERS
@nytai