@willbarrett 2020-04-28
I think Airbnb might have opinions about this. @john-bodley would this mess with your users' workflow?