@john-bodley 2020-04-28
@dpgaspar regarding the comment, 

> Since druid support outside of SQLAlchemy is deprecated

this isn't quite true, i.e., although we encourage environments to use Druid SQL (rather than Druid NoSQL) we still currently support Druid NoSQL and thus I'm not certain whether we should merge this PR.

I sense this PR shows the potential performance wins of actually fully deprecating the Druid NoSQL connector, i.e., there are numerous other places in the code base were the logic is complex and/or requires additional joins because there doesn't exist a foreign key between the `slices` and `tables` (or `datasources` which will be deprecated) tables.