@dpgaspar 2020-04-30
@john-bodley had the wrong impression regarding Druid NoSQL. 

Yet, adapted this PR so that we can have the best of both worlds. We still have the performance boost for fetching charts outside of Druid NoSQL and an extra query is issued for each Druid NoSQL existent on the query page (like before).

@willbarrett 
No user impact, just user satisfaction :)